requirements.txt
requests
phonenumbers
python-whois
colorama
Pillow
folium
fpdf2
dnspython
mac-vendor-lookup
pyinstaller
